+++ example/beaglebone-black/led.js
@@ -1,0 +1,15 @@
+'use strict';
+
+var five = require('johnny-five');
+var TinyBeagleBoneIO = require('./tiny-beaglebone-io');
+
+var board = new five.Board({
+  io: new TinyBeagleBoneIO()
+});
+
+board.on('ready', function() {
+  var led = new five.Led(46);
+
+  led.blink(500);
+});
+